requirements.txt
streamlit
openai
python-dotenv
openai-whisper
ffmpeg-python
torch

